Update index.html
Displays a pop-up box when a marker is detected and closes the pop-up box when the "Close" button is clicked. 

Defines showPopup function, with function call to showPopup placed inside the handleMarkerDetection function.

The pop-up box will show the corresponding image and information based on the object index of the detected marker.